Firther limits on number of types used in confined encoding
