+++ confined_encoding/Cargo.toml
@@ -21,7 +21,6 @@ crate-type = ["rlib", "staticlib"]
 [dependencies]
 amplify = { version = "4.0.0-alpha.6", features = ["proc_attr", "apfloat"] }
 confined_encoding_derive = { version = "0.9.0-alpha.1", path = "./derive", optional = true }
-bitcoin_hashes = "0.11.0" # We need this separately since bitcoin is an optional dependency
 bitcoin = "0.29.2"
 lnpbp_secp256k1zkp = { version = "0.9.0-alpha.1", optional = true }
 chrono = "0.4.23"
