+++ confined_encoding/src/bitcoin.rs
@@ -15,14 +15,44 @@
 use std::io;
 use std::io::{Read, Write};
 
+use bitcoin::hashes::{
+    hash160, hmac, ripemd160, sha256, sha256d, sha256t, sha512, Hash,
+};
 use bitcoin::util::taproot::{
     FutureLeafVersion, LeafVersion, TapBranchHash, TapLeafHash, TapTweakHash,
     TaprootMerkleBranch,
 };
 use bitcoin::{schnorr as bip340, secp256k1, OutPoint, Txid, XOnlyPublicKey};
-use bitcoin_hashes::sha256;
 
 use crate::{strategies, ConfinedDecode, ConfinedEncode, Error, Strategy};
+
+impl Strategy for sha256::Hash {
+    type Strategy = strategies::HashFixedBytes;
+}
+impl Strategy for sha256d::Hash {
+    type Strategy = strategies::HashFixedBytes;
+}
+impl<T> Strategy for sha256t::Hash<T>
+where
+    T: sha256t::Tag,
+{
+    type Strategy = strategies::HashFixedBytes;
+}
+impl Strategy for sha512::Hash {
+    type Strategy = strategies::HashFixedBytes;
+}
+impl Strategy for ripemd160::Hash {
+    type Strategy = strategies::HashFixedBytes;
+}
+impl Strategy for hash160::Hash {
+    type Strategy = strategies::HashFixedBytes;
+}
+impl<T> Strategy for hmac::Hmac<T>
+where
+    T: Hash,
+{
+    type Strategy = strategies::HashFixedBytes;
+}
 
 impl Strategy for Txid {
     type Strategy = strategies::HashFixedBytes;
